Refactored CreateTriangleGeometryAction to be template class.
This works for Triangle and Quad Geometries.
